Fix sound theme defaults for new users
frontend/common/src/SoundPlayer.cc

- (SoundPlayer::activate_theme)
Do not check that the filename is not empty before testing it. An empty
filename will now fail validation the same as if the filename does not
exist.

Prior to this change a new user would not receive the default settings
of a sound profile. Here's what happened with the break prelude sound,
for example:

A new user doesn't have the following keys:
HKEY_CURRENT_USER\AppEvents\Schemes\Apps\Workrave-crash\WorkraveBreakPrelude\.current
HKEY_CURRENT_USER\AppEvents\Schemes\Apps\Workrave-crash\WorkraveBreakPrelude\.default

So Workrave sets
HKEY_CURRENT_USER\Software\Workrave\sound\events\break_prelude as ""

Later activate_theme() is called by register_sound_events() and it
tests if break_prelude exists, and because it does it will not attempt
to set the filename. The filename, empty as "", will not fail
validation if it's empty.

As far as I know this was only a problem on Windows.