Change harpoonHelper project to use static CRT
common/win32/buildone.cmd
common/win32/harpoonHelper/src/harpoonHelper.vcproj

- (buildone.cmd)
Enable command processor extensions (ENABLEEXTENSIONS) and delayed
environment variable expansion (ENABLEDELAYEDEXPANSION).

On my systems with VS2008 and VS2010 I cannot run buildone.cmd
successfully without both of those options enabled. They are already
enabled in buildall.cmd.


- (harpoonHelper.vcproj)
Use the static release/debug CRT when building harpoonhelper.exe.

C++ code generation for the helper links to the release/debug CRT DLL
respectively depending on the configuration. A result of this is the
64-bit CRT DLL must be redistributed with Workrave because the version
required may not be on the user's computer already. That's a problem
because the helper is not in a separate directory where the 64-bit CRT
DLL can also be placed. The only CRT DLLs that should be placed in the
directory with workrave.exe are 32-bit types; there should be no 64-bit
types.

Other solutions would be move harpoonHelper.exe to a 64-bit directory
(eg something like %ProgramFiles%\Workrave\lib\x64) and include the
64-bit CRT DLL it was linked to, or include Microsoft's redist
installer for the 64-bit CRT DLL and have setup run it before install.

This is not currently a problem with the 64-bit applet or 64-bit
harpoon, both of which already link to the static CRT.