Add w32 topmost call in BreakWindow::refresh()
frontend/gtkmm/src/BreakWindow.cc

- (BreakWindow::refresh)
Add a call to SetWindowPos() with HWND_TOPMOST to ensure that our break
window stays topmost even over other topmost windows.

Prior to this change the call to W32Compat::SetWindowOnTop() --which
is similar to what is described above-- in BreakWindow::refresh() was
eliminated for compatibility with W32 force focus behavior.
Ref: e5df39fb31b10745697406ebe2dc7c4283a75d05

All of the behavior of W32Compat::SetWindowOnTop() is now in
BreakWindow::refresh() and is modified to be force focus compatible.