fix memory leak when play sound.
I find this issue for very long time.
However, I didn't find a good way to fix it.
Now, I use a bad way to do this.

As QTKIT/QUICKTIME is depreceated in OSX, I change to NSSound.
I keep an dictionary for the sound file and NSSound object to avoid
new instance for NSSound every time.