New timer icon set
These icons are optimized for 16x16 usage, so be careful about using
them in other contexts.

The included .afdesign files can be opened with Affinity Designer.  They
were saved in a lossless vector format. I originally tried exporting the
graphics to SVG and EPS since they're non-proprietary. Unfortunately,
both exports were lossy due to rasterization so this is a compromise to
preserve the original quality of the vector files.