Fix open_window on system tray reloads
Fixes #612

Short unavailabilities of the system tray cause the status window to appear.
This can happen, when Sway reloads or shortly after waking a monitor from
sleep.
The window has to be closed, which is cumbersome.

This patch fixes this by introducing a short delay to showing the window,
which gives the system tray time to reappear.
It also checks, if there is any monitor to show the window on, before
showing it, mainly for not setting the status window setting to enabled,
which causes it to reappear after waking up a monitor.